# Speed up eval of log_prob for InverseWishartPrior