@jacobrgardner 2018-07-20
I'll merge this in. If we ever have larger or structured matrices (e.g., low rank for ICM kernel), we'll probably want to consider making use of a `LazyVariable.inv_matmul` to do this with CG.